+++ core/dialog/prompt.svelte
@@ -7,7 +7,7 @@
     import Paper from "../paper.svelte"
     import Text from "../text.svelte"
     import TextInput from "../text-input.svelte"
-    import TitleBar from "../title-bar.svelte"
+    import Titlebar from "../titlebar.svelte"
 
     import DialogContent from "./content.svelte"
 
@@ -45,12 +45,12 @@
     <Paper card>
         <svelte:fragment slot="title">
             {#if title}
-                <TitleBar compact>
+                <Titlebar compact>
                     {#if icon}
                         <Icon name={icon} />
                     {/if}
                     {title}
-                </TitleBar>
+                </Titlebar>
             {/if}
         </svelte:fragment>
         <Text>