+++ core/select.svelte
@@ -1,11 +1,10 @@
 <script>
-    import Adornment from "./adornment.svelte"
     import Button from "./button.svelte"
     import ControlDrawer from "./control-drawer.svelte"
     import Icon from "./icon.svelte"
     import Text from "./text.svelte"
     import TextInput from "./text-input.svelte"
-    import TitleBar from "./title-bar.svelte"
+    import Titlebar from "./titlebar.svelte"
 
     import OptionList from "./select/option-list.svelte"
 
@@ -23,6 +22,7 @@
         filter = ""
         value = newValue
     }
+    const open = () => drawer.open()
 
     let filter = ""
 
@@ -43,35 +43,64 @@
     select-layout {
         display: grid;
         flex-grow: 1;
-        grid-template-columns: auto max-content;
+        grid-template-columns: 24px auto;
+        grid-template-rows: min-content auto;
+    }
+
+    label-area {
+        display: block;
+        user-select: none;
+        font-size: 13px;
+        grid-column: span 2;
+        padding: 2px 16px;
+        border-bottom-right-radius: 4px;
+        width: max-content;
+        cursor: default;
+
+        border-bottom: 1px solid var(--control-border);
+        border-right: 1px solid var(--control-border);
+
+        color: var(--control-border);
+    }
+    label-area:empty {
+        display: none;
+    }
+
+    selected-area {
+        padding: 8px;
     }
 </style>
 
 <ControlDrawer {persistent} bind:this={drawer}>
     {#if label}
-        <TitleBar compact sticky>
+        <Titlebar compact sticky>
             {label}
-            <Adornment slot="extension">
+            <svelte:fragment slot="extension">
                 {#if searchable}
-                    <TextInput bind:value={filter} label="Filter" />
+                    <TextInput bind:value={filter} label="Filter" adorn />
                 {/if}
-            </Adornment>
-        </TitleBar>
+            </svelte:fragment>
+        </Titlebar>
     {/if}
     <slot name="options" {info}>
         <OptionList {info} />
     </slot>
 </ControlDrawer>
 
-<Button variant="outline" {...$$props} on:click={() => drawer.open()} {disabled}>
+<Button variant="outline" {...$$props} on:click={open} {disabled} control>
     <select-layout>
-        <Text adorn>
-            <slot name="selected" {selected}>
-                {label}: {selected.label}
-            </slot>
-        </Text>
+        <label-area>
+            {label}
+        </label-area>
         <Text adorn>
             <Icon name={icon} />
         </Text>
+        <selected-area>
+            <Text>
+                <slot name="selected" {selected}>
+                    {selected?.label ?? "Not Selected"}
+                </slot>
+            </Text>
+        </selected-area>
     </select-layout>
 </Button>