+++ site/src/component/adornment.svelte
@@ -1,15 +1,10 @@
 <script>
-    import {
-        Avatar,
-        TitleBar,
-    } from "@core"
-
     import Markdown from "../app/markdown.svelte"
     import docs from "@docs"
 </script>
 
 <h1>
-    Adornment
+    Adornments
 </h1>
 
 <Markdown {docs} />