+++ site/src/component/baseline.svelte
@@ -1,8 +1,4 @@
 <script>
-    import {
-        TitleBar,
-    } from "@core"
-
     import Markdown from "../app/markdown.svelte"
     import docs from "@docs"
 </script>